+++ allensdk/__init__.py
@@ -38,7 +38,7 @@ import logging
 
 
 
-__version__ = '2.6.0'
+__version__ = '2.7.0'
 
 
 try: